ref(statistics): rename utils package for clarity
