+++ statistics/stats_test.go
@@ -1,4 +1,4 @@
-package utils
+package statistics
 
 import (
 	"reflect"
